+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDProfileId.cs
@@ -139,7 +139,6 @@ namespace UnityEngine.Rendering.HighDefinition
         RenderDeferredLightingSinglePassMRT,
 
         // Misc
-        BuildMarginals,
         VolumeUpdate,
         CustomPassVolumeUpdate,
         XROcclusionMesh,